requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.27,<0.35
psycopg2-binary>=2.9,<3.0
numpy>=1.24,<3.0
pydantic>=2.5,<3.0
